fi:find-tag -> fi:lisp-find-definition and fi:lisp-tags-loop-continue -> fi:lisp-find-next-definition. Furthermore, fi:lisp-find-definition is on ``C-c .'' and fi:lisp-find-next-definition is on ``C-c ,''.  ``C-c 4 .'' is bound to fi:lisp-find-definition-other-window.
Implement abbreviations (for example, f-d-c-p expands to
fast-digit-char-p).

``C-c SPC'' restores the window configuration as it was before the
temporary window was made visible.

M-. is locked for first-time use, displaying a message about the
new changes.  New variable fi:find-tag-lock.

Make the second argument to interrupt-process, kill-process and
quit-process depend on the mode--in rlogin, shell and telnet modes
it is `t', otherwise `nil'