bug25315: ELI fixes for Emacs 26.1
ELI was behaving strangely on 26.1.  The strategy was to first get rid
of build warnings and hope for the best.  Seems to have worked.

Also remove support for Emacs 18 and 19.

The only remaining warnings are due to these:

  Warning: global/dynamic var `session' lacks a prefix
  Warning: global/dynamic var `connection' lacks a prefix
  Warning: global/dynamic var `process' lacks a prefix

Fixing these will require careful study, as I believe these variables
are used dynamically, and there are many decoys in the source code.

Change-Id: I465df2dea28381e62bd3c8b06e2361ecd78fd88a
Reviewed-on: https://gerrit.franz.com:9080/10929
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>