Workaround 28.1 bug
Change the defun of common-lisp-mode to a defalias.  This is both a
workaround and the best way to replace common-lisp-mode, since
originally common-lisp-mode is an alias for lisp-mode.

The test case is to visit a file with `mode: common-lisp` in the mode
line.  Without this change, in 28.1, there will be infinite recursion
in run-mode-hooks.

%Testing: manual

Change-Id: I4179268bc2e407aba171055169611f336cf5de9e
Reviewed-on: https://gerrit.franz.com:9080/c/eli/+/20511
Tested-by: Kevin Layer <layer@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>